requirements.txt
google_api_python_client==2.66.0
google_auth_oauthlib==0.4.6
nltk==3.7
numpy==1.19.5
openai==0.25.0
protobuf==3.20.*
pyttsx3==2.90
pytz==2022.5
SpeechRecognition==3.8.1
tensorflow==2.5.0
torch==1.13.0
transformers==4.24.0
PyAudio
psycopg2
python-dotenv
wikipedia
playsound
pyowm
pyjokes
sklearn
